HW-57735 move TextWatcher impl base class
